requirements.txt
numpy==1.20.1
pandas==1.2.4
Pillow==8.4.0
torch==1.10.0
torchvision==0.11.1
tqdm==4.59.0
